requirements.txt
streamlit
aiomysql
python-dotenv
ollama
aiohttp
